+++ .env
@@ -1,2 +1,7 @@
 # .env
-DISCORD_TOKEN = ###TOKEN HERE###
+
+DISCORD_TOKEN = ### TOKEN HERE ###
+
+BACKGROUND_LOOP_TIME = 60  # seconds of running loop for party cleanup
+COMMAND_PREFIX = '.'
+LFG_CHANNEL = 'lookingforgroup'
